Refactor custom logic into new files, for less conflicts when we update CometBFT version.
